+++ config/initializers/optimacms.rb
@@ -1,7 +1,6 @@
-Optimacms.main_namespace = ''
-Optimacms.admin_namespace = 'admin'
-Optimacms.files_dir_path = 'img'
+Optimacms.config do |c|
+  c.main_namespace = ''
+  c.admin_namespace = 'admin'
+  c.files_dir_path = 'img'
+end
 
-Optimacms.backup_metadata = {
-    dir_base: "/data/backups/tpl/metadata"
-}