+++ changelog.md
@@ -1,3 +1,20 @@
+## 2018-april
+
+* new config for Optimacms with config/optimacms.yml
+
+* added Options to Admin Area
+
+
+## 2018-march
+
+
+* font-awesome
+* refactoring CSS
+
+* app data settings
+
+
+
 ## 2017-nov
 
 * ruby 2.4.2